+++ .prettierignore
@@ -1,14 +1,21 @@
 **/fixtures/**/*.js
+**/fixtures/**/*.json
 **/fixtures/**/*.ts
 **/template/**/*.js
 **/template/**/*.ts
 **/templates/**/*.js
 **/templates/**/*.ts
-**/bin/*.js
+**/templates/**/*.json
 **/website/*.js
-packages/umi-build-dev/src/plugins/commands/block/index.js
 packages/af-webpack/src/getConfig/index.js
-packages/af-webpack/src/getConfig/css.js
-packages/af-webpack/src/getConfig/eslint.js
-packages/af-webpack/src/getConfig/dev.js
+packages/af-webpack/src/getConfig/css.js	
+packages/af-webpack/src/getConfig/eslint.js	
+packages/af-webpack/src/getConfig/dev.js	
 packages/af-webpack/src/getConfig/prod.js
+**/*.svg
+package.json
+.umi
+.umi-production
+/dist
+.dockerignore
+.eslintignore
