+++ .prettierrc
@@ -1,5 +1,10 @@
 {
   "printWidth": 100,
   "singleQuote": true,
-  "trailingComma": "all"
+  "trailingComma": "all",
+  "proseWrap": "never",
+  "overrides": [
+    { "files": ".webpackrc", "options": { "parser": "json" } },
+    { "files": ".prettierrc", "options": { "parser": "json" } }
+  ]
 }
