+++ .travis.yml
@@ -1,8 +1,8 @@
 language: node_js
 
 node_js:
-  - "8"
-  - "10"
+  - '8'
+  - '10'
 
 before_script:
   - npm run build
