+++ package.json
@@ -5,20 +5,21 @@
     "build": "father build",
     "changelog": "lerna-changelog",
     "chore:update-deps": "./scripts/update_deps.sh",
-    "test": "node scripts/test.js",
-    "test:coverage": "node ./packages/umi-test/bin/umi-test.js --coverage",
-    "debug": "./packages/umi-test/bin/umi-test.js",
     "coveralls": "cat ./coverage/lcov.info | coveralls",
+    "debug": "./packages/umi-test/bin/umi-test.js",
+    "doc:deploy": "rm -rf ./website/yarn.lock && cd ./website && npm run deploy && cd -",
+    "doc:dev": "./website/node_modules/.bin/vuepress dev ./docs",
     "lint": "eslint --ext .js packages",
     "precommit": "lint-staged",
-    "doc:dev": "./website/node_modules/.bin/vuepress dev ./docs",
-    "doc:deploy": "rm -rf ./website/yarn.lock && cd ./website && npm run deploy && cd -",
+    "prettier": "prettier --write '**/*.{js,jsx,tsx,ts,less,md,json}'",
     "publish": "./scripts/publish.js",
-    "ui:dev": "APP_ROOT=packages/umi-build-dev/src/plugins/commands/ui ./packages/umi/bin/umi.js dev",
-    "ui:build": "APP_ROOT=packages/umi-build-dev/src/plugins/commands/ui ./packages/umi/bin/umi.js build"
+    "test": "node scripts/test.js",
+    "test:coverage": "node ./packages/umi-test/bin/umi-test.js --coverage",
+    "ui:build": "APP_ROOT=packages/umi-build-dev/src/plugins/commands/ui ./packages/umi/bin/umi.js build",
+    "ui:dev": "APP_ROOT=packages/umi-build-dev/src/plugins/commands/ui ./packages/umi/bin/umi.js dev"
   },
   "lint-staged": {
-    "*.(t|j)s": [
+    "**/*.{js,jsx,tsx,ts,less,md,json}": [
       "prettier --write",
       "git add"
     ]
