+++ packages/umi-build-dev/src/plugins/commands/ui/index-debug.html
@@ -1,29 +1,32 @@
-<!DOCTYPE html><html><head>
+<!DOCTYPE html>
+<html>
+  <head>
+    <link rel="stylesheet" href="https://unpkg.com/antd@3.11.0/dist/antd.min.css" />
+    <link rel="stylesheet" href="http://localhost:8002/umi.css" />
 
-  <link rel="stylesheet" href="https://unpkg.com/antd@3.11.0/dist/antd.min.css">
-  <link rel="stylesheet" href="http://localhost:8002/umi.css">
+    <meta charset="utf-8" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1, maximum-scale=1, minimum-scale=1, user-scalable=no"
+    />
+    <title></title>
+    <script>
+      window.routerBase = '/';
+    </script>
+    <script>
+      window.g_umiUIPlugins = [];
+    </script>
+  </head>
+  <body>
+    <script src="https://unpkg.com/react@16.7.0-alpha.2/umd/react.development.js"></script>
+    <script src="https://unpkg.com/react-dom@16.7.0-alpha.2/umd/react-dom.development.js"></script>
+    <script src="https://unpkg.com/moment@2.22.2/min/moment.min.js"></script>
+    <script src="https://unpkg.com/antd@3.11.0/dist/antd.min.js"></script>
+    <script src="https://unpkg.com/dva@2.4.1/dist/dva.min.js"></script>
+    <script src="https://unpkg.com/sockjs-client@1.3.0/dist/sockjs.min.js"></script>
 
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, minimum-scale=1, user-scalable=no">
-  <title></title>
-  <script>
-    window.routerBase = "/";
-  </script>
-  <script>
-    window.g_umiUIPlugins = [];
-  </script>
-</head>
-<body>
+    <div id="root"></div>
 
-<script src="https://unpkg.com/react@16.7.0-alpha.2/umd/react.development.js"></script>
-<script src="https://unpkg.com/react-dom@16.7.0-alpha.2/umd/react-dom.development.js"></script>
-<script src="https://unpkg.com/moment@2.22.2/min/moment.min.js"></script>
-<script src="https://unpkg.com/antd@3.11.0/dist/antd.min.js"></script>
-<script src="https://unpkg.com/dva@2.4.1/dist/dva.min.js"></script>
-<script src="https://unpkg.com/sockjs-client@1.3.0/dist/sockjs.min.js"></script>
-
-<div id="root"></div>
-
-<script src="http://localhost:8002/umi.js"></script>
-
-</body></html>
+    <script src="http://localhost:8002/umi.js"></script>
+  </body>
+</html>
