+++ packages/umi-plugin-hd/README.md
@@ -5,4 +5,3 @@ Hd plugin for umi.
 ---
 
 Suggest to use together with umi-plugin-react, see our website [umi-plugin-react](https://umijs.org/plugin/umi-plugin-react.html) for more.
-
