+++ packages/umi-plugin-locale/README.md
@@ -21,16 +21,16 @@ export default {
         locale: {
           default: 'zh-CN', //默认语言 zh-CN
           baseNavigator: true, // 为true时，用navigator.language的值作为默认语言
-          antd: true // 是否启用antd的<LocaleProvider />
-        }
-      }
-    ]
-  ]
+          antd: true, // 是否启用antd的<LocaleProvider />
+        },
+      },
+    ],
+  ],
 };
 ```
 
-Other:
-You can setDefaultLocal in `src/app.js`
+Other: You can setDefaultLocal in `src/app.js`
+
 ```js
 export const locale = {
   default: 'en-US', //默认语言 zh-CN
@@ -41,27 +41,24 @@ export const locale = {
 
 ```
 .
-├── dist/                          
-├── mock/                         
-└── src/                          
-    ├── layouts/index.js          
-    ├── pages/                    
+├── dist/
+├── mock/
+└── src/
+    ├── layouts/index.js
+    ├── pages/
     └── locales               // 多语言文件存放目录，里面的文件会被umi自动读取
         ├── zh-CN.js
-        └── en-US.js               
-├── .umirc.js                     
-├── .env                          
+        └── en-US.js
+├── .umirc.js
+├── .env
 └── package.json
 ```
 
-
->如果`.umirc.js`里配置了`singular: true`，`locales`要改成`locale`
-
+> 如果`.umirc.js`里配置了`singular: true`，`locales`要改成`locale`
 
 ## 多语言文件约定
 
 多语言文件的命名规范：`<lang>-<COUNTRY>.js`
-
 
 多语言文件的内容规范：键-值组成的字面量，如下：
 
@@ -70,18 +67,18 @@ zh-CN.js
 ```javascript
 export default {
   WELCOME_TO_UMI_WORLD: '{name}，欢迎光临umi的世界',
-}
+};
 ```
 
 en-US.js
 
 ```javascript
 export default {
-  WELCOME_TO_UMI_WORLD: '{name}, welcome to umi\'s world',
-}
+  WELCOME_TO_UMI_WORLD: "{name}, welcome to umi's world",
+};
 ```
 
-## 扩展API
+## 扩展 API
 
 使用本插件后，你可以从 umi-plugin-locale 引入国际化相关的 API，获得关于多语言功能在编程上的便利。
 
@@ -89,11 +86,14 @@ export default {
 import { formatMessage, setLocale, getLocale, FormattedMessage } from 'umi-plugin-locale';
 
 // 获取指定文字的多语言版本
-const formatedText = formatMessage({
-  id: 'WELCOME_TO_UMI_WORLD'
-}, {
-  name: '小伙子',
-});
+const formatedText = formatMessage(
+  {
+    id: 'WELCOME_TO_UMI_WORLD',
+  },
+  {
+    name: '小伙子',
+  },
+);
 
 console.log(formatedText === '小伙子，欢迎光临umi的世界');
 
@@ -103,10 +103,8 @@ setLocale('en-US');
 // 获取当前语言
 console.log(getLocale() === 'en-US');
 
-
 // 渲染一个文字标签
 function Example() {
   return <FormattedMessage id="WELCOME_TO_UMI_WORLD" values={{ name: '小伙子' }} />;
 }
 ```
-
