+++ packages/umi-plugin-modern-mode/README.md
@@ -1,7 +1,6 @@
 # umi-plugin-umi-plugin-modern-mode
 
-[![NPM version](https://img.shields.io/npm/v/umi-plugin-umi-plugin-modern-mode.svg?style=flat)](https://npmjs.org/package/umi-plugin-umi-plugin-modern-mode)
-[![NPM downloads](http://img.shields.io/npm/dm/umi-plugin-umi-plugin-modern-mode.svg?style=flat)](https://npmjs.org/package/umi-plugin-umi-plugin-modern-mode)
+[![NPM version](https://img.shields.io/npm/v/umi-plugin-umi-plugin-modern-mode.svg?style=flat)](https://npmjs.org/package/umi-plugin-umi-plugin-modern-mode) [![NPM downloads](http://img.shields.io/npm/dm/umi-plugin-umi-plugin-modern-mode.svg?style=flat)](https://npmjs.org/package/umi-plugin-umi-plugin-modern-mode)
 
 A plugin for modern mode
 
@@ -11,10 +10,8 @@ Configure in `.umirc.js`,
 
 ```js
 export default {
-  plugins: [
-    ['umi-plugin-umi-plugin-modern-mode', options],
-  ],
-}
+  plugins: [['umi-plugin-umi-plugin-modern-mode', options]],
+};
 ```
 
 ## Options
@@ -22,6 +19,7 @@ export default {
 ```
 { isModernBuild = true, unsafeInline = false } = Options
 ```
+
 isModernBuild: modern mode
 
 unsafeInline: inline safari nomodule pollyfill
